Move opencv-341.jar to lib folder
I searched in a lot of places for how to import .jar files in VSCode,
and did not have much success. It seems VSCode prefers to use a project
build file (Maven, Gradle) to read dependencies. However, there's a new
feature where you put .jar files in a lib folder and it works.

https://github.com/redhat-developer/vscode-java/issues/384
https://github.com/Microsoft/vscode-java-debug/issues/194
https://github.com/redhat-developer/vscode-java/issues/501
https://github.com/eclipse/eclipse.jdt.ls/pull/927