Add error correction when a corner becomes empty but the system does not recognize it
The last added manual checking, to see if a now empty corner is valid, makes
some mistakes when the board moves at the same time a corner stone is removed.
This is not a common situation, but the error correction of this commit
corrects it. This situation can be found on frame 93 of sequence 17.